backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.29
gunicorn==21.2.0
psycopg2-binary==2.9.9  # PostgreSQL driver
gevent==23.9.1  # For Gunicorn async workers